changed switch statements to array
Previously, this program used switch statements. as time goes by, that
may cause bloating, which is why I changed it to read strings from an
array. Maybe in the future I will implement another example that uses
Sets, and create a second program that allows you to add fortunes
without having to modify the code.